# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "joinedcontext Portal API",
    "description": "Administrative and platform management REST API for joinedcontext Portal",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project}/syncsources/{name}/detach": {
      "post": {
        "tags": [
          "resources"
        ],
        "operationId": "detach",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "SyncSource name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "The merge request that removes the source",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such project or source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No repository configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/syncsources/{name}/pause": {
      "post": {
        "tags": [
          "resources"
        ],
        "operationId": "pause",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "SyncSource name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "What the source reports afterwards",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncSourceStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such project or source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No repository configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/syncsources/{name}/status": {
      "get": {
        "tags": [
          "resources"
        ],
        "operationId": "status",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "SyncSource name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "What the source reports about itself",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncSourceStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such project or source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No repository configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/syncsources/{name}/sync": {
      "post": {
        "tags": [
          "resources"
        ],
        "operationId": "sync_now",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "SyncSource name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run and what the source reports afterwards",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunReport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such project or source",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No repository configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/webhooks/sync/{project}/{name}": {
      "post": {
        "tags": [
          "system"
        ],
        "operationId": "webhook",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "SyncSource name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-gitea-signature",
            "in": "header",
            "description": "HMAC-SHA256 of the request body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Whatever the origin sends; the body is what the signature covers",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The run the source's origin asked for",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunReport"
                }
              }
            }
          },
          "401": {
            "description": "The signature is not one this source's own secret makes over this body — the same answer as for a source that is not there, one with no `spec.webhook`, and one whose reference this instance cannot resolve (MF-44, PF-59)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No repository configured, or no sync loop running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PauseRequest": {
        "type": "object",
        "description": "Whether syncing is on or off.",
        "required": [
          "paused"
        ],
        "properties": {
          "paused": {
            "type": "boolean",
            "description": "`true` switches the loop off, `false` switches it back on."
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details representation.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Every violation of one request, when there is more than one thing to say. A form marks\nall its bad fields from this in one pass instead of one round trip per mistake (CC-24)."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "SyncRunReport": {
        "type": "object",
        "description": "What one run did, and where the source stands afterwards.",
        "required": [
          "proposed",
          "unchanged",
          "flags",
          "status"
        ],
        "properties": {
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "What the run could not do, in sentences a person can act on."
          },
          "proposed": {
            "type": "array",
            "items": {},
            "description": "The `kind: Change` envelope of every merge request the run opened."
          },
          "status": {
            "$ref": "#/components/schemas/SyncSourceStatus"
          },
          "unchanged": {
            "type": "integer",
            "description": "Whether the run changed anything at all.",
            "minimum": 0
          }
        }
      },
      "SyncSourceStatus": {
        "type": "object",
        "description": "What the project page shows for one source (MF-30).",
        "required": [
          "project",
          "name",
          "phase",
          "paused",
          "durable"
        ],
        "properties": {
          "durable": {
            "type": "boolean",
            "description": "Whether a restart keeps this. `false` on a Portal without a database, where a restart\ncosts one duplicate proposal per source with a run in flight."
          },
          "lastError": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the last run did not finish, when it did not."
          },
          "lastRunAt": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "When the last run happened, in seconds since the epoch.",
            "minimum": 0
          },
          "mergeRequest": {
            "type": [
              "string",
              "null"
            ],
            "description": "The merge request a run opened and nobody has answered yet."
          },
          "name": {
            "type": "string"
          },
          "observedRevision": {
            "type": [
              "string",
              "null"
            ],
            "description": "The source revision the repository carries."
          },
          "paused": {
            "type": "boolean"
          },
          "phase": {
            "type": "string",
            "description": "`Synced`, `OutOfSync`, `PendingApproval`, `Error` or `Paused`."
          },
          "project": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}